Move the get_user_auth_backend method from resources app to users
Move the method that extracts the user's login backend from
resources app to users app as this is a better place.

Refs None
